create questions and teams folder
