+++ client/src/scenes/Game.ts
@@ -8,10 +8,10 @@ import GameUi from "~/scenes/GameUi";
 import Lizard from "~/enemies/Lizard";
 import * as Colyseus from "colyseus.js";
 import {
-    SetupPlayerAnimsUpdate,
-    SetupPlayerOnCreate,
-    SetUpPlayerSyncWithServer,
-    SetUpPlayerListeners
+  SetupPlayerAnimsUpdate,
+  SetupPlayerOnCreate,
+  SetUpPlayerSyncWithServer,
+  SetUpPlayerListeners
 } from "~/anims/PlayerSync";
 import { ButtonCreator } from "~/components/ButtonCreator";
 import { setUpVoiceComm } from "~/communications/SceneCommunication";
@@ -91,10 +91,10 @@ export default class Game extends Phaser.Scene {
 
       this.collisionSetUp();
 
-            SetUpPlayerListeners(this);
-        } catch (e) {
-            console.error("join error", e);
-        }
+      SetUpPlayerListeners(this);
+    } catch (e) {
+      console.error("join error", e);
+    }
 
     try {
       this.setBattleQueueInteractiveUi();
@@ -282,10 +282,10 @@ export default class Game extends Phaser.Scene {
       "In Queue: " +
       (this.queueList.length > 0
         ? this.queueList
-            .map((userName) =>
-              userName === this.currentUsername ? "Me" : userName,
-            )
-            .join(", ")
+          .map((userName) =>
+            userName === this.currentUsername ? "Me" : userName,
+          )
+          .join(", ")
         : "No players");
 
     if (!this.queueDisplay) {
